% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compositionJF}
\alias{compositionJF}
\title{Composition plot for Joes Flow}
\usage{
compositionJF(meta, grp, kmeans_groups, colors)
}
\arguments{
\item{meta}{Data frame containing metadata labels. Names of each element of the vector should correspond to sample IDs and include all IDs in `kmeans_groups`.}

\item{grp}{Column of meta to use for grouping of data}

\item{kmeans_groups}{Named character vector containing Kmeans group label for each sample. Names of each element of the vector should correspond to sample IDs and include all IDs in `meta`.}

\item{colors}{Vector of colors for clusters}
}
\value{
A ggplot object
}
\description{
Composition plot for Joes Flow
}
