% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dimreductJF}
\alias{dimreductJF}
\alias{dimreductJF.prcomp}
\alias{dimreductJF.matrix}
\title{Plot a dimension reduction figure for JoesFlow}
\usage{
dimreductJF(x, sample_data, features, ...)

\method{dimreductJF}{prcomp}(x, sample_data, features, ...)

\method{dimreductJF}{matrix}(x, sample_data, features, ...)
}
\arguments{
\item{x}{Object containing clustered data (expects output from `prcomp`, `uamp`, or `Rtsne`)}

\item{sample_data}{sample data}

\item{features}{named vector, features in names(sample_data), names(features) = unique(kmeans_groups) -- names should be ordered}

\item{...}{Other objects passed along to functions within clusterJF}
}
\value{
A ggplot object
}
\description{
Plot a dimension reduction figure for JoesFlow
}
