% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_values.R
\name{extract_sb_loadings}
\alias{extract_sb_loadings}
\title{extract_sb_loadings
Extract loadings from sample based PCA}
\usage{
extract_sb_loadings(clustered_data)
}
\arguments{
\item{clustered_data}{Object containing clustered data (expects output from `prcomp`)}
}
\value{
a data frame with loadings for PC1 and PC2 for each Cluster
}
\description{
extract_sb_loadings
Extract loadings from sample based PCA
}
