% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_values.R
\name{extract_sb_values}
\alias{extract_sb_values}
\title{extract_sb_values
Extract values from sample based PCA}
\usage{
extract_sb_values(clustered_data, ids, meta, grp, cluster = NULL)
}
\arguments{
\item{clustered_data}{Object containing clustered data (expects output from `prcomp`)}

\item{ids}{Character vector of ids for each row in `clustered_data$x`, corresponding to labels in `grps`}

\item{meta}{Data frame containing translation from id to group}

\item{grp}{Character value identifying the column of `meta` to use for group identifier}

\item{cluster}{Data frame containing sample ID and the assigned kmeans cluster, as returned by `kmeaner()`}
}
\value{
a data frame with values for SampleID, Group, PC1, and PC2
}
\description{
extract_sb_values
Extract values from sample based PCA
}
