% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_values.R
\name{extract_values}
\alias{extract_values}
\alias{extract_values.prcomp}
\alias{extract_values.matrix}
\title{extract_values
Extract values from PCA, UMAP, and tSNE}
\usage{
extract_values(clustered_data, ...)

\method{extract_values}{prcomp}(clustered_data, ids, meta, grp, cluster = NULL, ...)

\method{extract_values}{matrix}(clustered_data, ids, meta, grp, cluster = NULL, ...)
}
\arguments{
\item{clustered_data}{Object containing clustered data (expects output from `prcomp`, `umap`, or `tsne`)}

\item{...}{Other objects passed to methods of `extract_values`}

\item{ids}{Character vector of ids for each row in `clustered_data`, corresponding to labels in `grps`}

\item{meta}{Data frame containing translation from id to group}

\item{grp}{Character value identifying the column of `meta` to use for group identifier}

\item{cluster}{Data frame containing sample ID and the assigned kmeans cluster, as returned by `kmeaner()`}
}
\value{
A tibble with values for SampleID, Group, Cluster, PC/vector 1, and PC/vector 2
}
\description{
extract_values
Extract values from PCA, UMAP, and tSNE
}
