% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{marker_heatJF}
\alias{marker_heatJF}
\title{Marker heat plot for Joes Flow}
\usage{
marker_heatJF(sample_data, ids, meta, kmeans_groups, colors, sample_size)
}
\arguments{
\item{sample_data}{Data frame or numeric matrix containing sample data}

\item{ids}{Character vector of ids for each row in `sample_data`, corresponding to the labels in `meta`}

\item{meta}{Character vector containing metadata labels, corresponding to ids (when ids are sorted using `order`)}

\item{kmeans_groups}{Vector containing Kmeans group label for each row in `sample_data`}

\item{colors}{Vector of colors for samples}

\item{sample_size}{An integer indicating how many samples to use in the generation of the figure}
}
\value{
A grid graphics object
}
\description{
Marker heat plot for Joes Flow
}
