% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_data}
\alias{meta_data}
\title{Meta data}
\format{
A data frame with unique sample ID in the first column corresponding to the sample IDs in `sample_data` and meta data for each sample in subsequent columns (e.g. group, treatment, ...).
}
\usage{
meta_data
}
\description{
Example meta data
}
\keyword{datasets}
