% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run a local instance of JoesFlow}
\usage{
run_app()
}
\description{
This function will start up a local instance of JoesFlow from within RStudio.
}
\details{
While this is useful, the system requirements can take some time to set up properly. We recommend the use of Docker (see https://github.com/IDSS-NIAID/JoesFlow for more documentation on using Docker).
}
