% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample data}
\format{
A data frame with sample ID in the first column and flow metrics in each subsequent column. Each row represents a different cell.
}
\usage{
sample_data
}
\description{
Example sample data
}
\keyword{datasets}
