% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sb_clusterJF}
\alias{sb_clusterJF}
\title{Sample-based PCA for JoesFlow}
\usage{
sb_clusterJF(
  clustered_data,
  ids,
  meta,
  grp,
  colors1,
  colors2,
  legend.name = "Group",
  show_grp_legend = TRUE,
  show_clust_legend = TRUE
)
}
\arguments{
\item{clustered_data}{Object containing clustered data (expects output from `prcomp` or `umap`)}

\item{ids}{Character vector of ids for each row in `clustered_data$x`, corresponding to labels in `meta`}

\item{meta}{Data frame containing metadata - first column corresponds to ids}

\item{grp}{Column of meta to use for grouping of data}

\item{colors1}{Vector of colors for samples}

\item{colors2}{Vector of colors for clusters}

\item{legend.name}{Character string for the legend name (default is 'Group')}

\item{show_grp_legend}{Logical value used to turn off the legend for groups}

\item{show_clust_legend}{Logical value used to turn off the legend for clusters}
}
\value{
A ggplot object
}
\description{
Sample-based PCA for JoesFlow
}
