% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sb_loadingsJF}
\alias{sb_loadingsJF}
\title{Plot loadings for a sample based cluster for JoesFlow}
\usage{
sb_loadingsJF(clustered_data, colors)
}
\arguments{
\item{clustered_data}{Object containing clustered data (expects output from `prcomp` or `uamp`)}

\item{colors}{Vector of colors for each group}
}
\value{
A ggplot object
}
\description{
Plot loadings for a sample based cluster for JoesFlow
}
