% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_utils.R
\name{setup_testing_data}
\alias{setup_testing_data}
\title{setup_testing_data
Set up testing data and return a list of files to test with}
\usage{
setup_testing_data()
}
\value{
Returns a character vector containing paths to files that need to be tested
}
\description{
This sets up testing data and returns a list of files to test with
}
\details{
To add other data not included as part of the repository, create a directory at root/testing/testData/ with the desired files. If that directory is missing, testing will only be performed on internal data (i.e. data included with the package).
}
