% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clusterJF}
\alias{clusterJF}
\title{Plot a cluster for JoesFlow}
\usage{
clusterJF(clustered_data, ids, meta, colors, legend.name = "Group")
}
\arguments{
\item{clustered_data}{Object containing clustered data (expects output from `prcomp` or `uamp`)}

\item{ids}{Character vector of ids for each row in `clustered_data$x`, corresponding to SampleID in `meta`}

\item{meta}{Data frame containing metadata, with SampleID in the first column and Group label in the second column}

\item{colors}{Vector of colors for each group}

\item{legend.name}{Character string for the legend name (default is 'Group')}
}
\value{
A ggplot object
}
\description{
Plot a cluster for JoesFlow
}
