% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_utils.R
\name{save_flowdata}
\alias{save_flowdata}
\title{save_flowdata
Save flow data}
\usage{
save_flowdata(flow, meta, RData)
}
\arguments{
\item{flow}{Path to flow data}

\item{meta}{Path to metadata}

\item{Rdata}{Path to output file}
}
\value{
Returns `TRUE` when successful.
}
\description{
Thie saves flow data and metadata to a .RData file for use in testing scripts
}
