% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sb_clusterJF}
\alias{sb_clusterJF}
\title{Sample-based PCA for JoesFlow}
\usage{
sb_clusterJF(
  clustered_data,
  ids,
  meta,
  colors1,
  colors2,
  legend.name = "Group"
)
}
\arguments{
\item{clustered_data}{Object containing clustered data (expects output from `prcomp` or `umap`)}

\item{ids}{Character vector of ids for each row in `clustered_data$x`, corresponding to labels in `meta`}

\item{meta}{Character vector containing metadata labels, corresponding to ids}

\item{colors1}{Vector of colors for samples}

\item{colors2}{Vector of colors for clusters}

\item{legend.name}{Character string for the legend name (default is 'Group')}
}
\value{
A ggplot object
}
\description{
Sample-based PCA for JoesFlow
}
