% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sb_clusterJF}
\alias{sb_clusterJF}
\title{Plot a sample based cluster for JoesFlow}
\usage{
sb_clusterJF(clustered_data, ids, meta, colors1, colors2)
}
\arguments{
\item{clustered_data}{Object containing clustered data (expects output from `prcomp` or `uamp`)}

\item{ids}{Character vector of ids for each row in `clustered_data$x`, corresponding to SampleID in `meta`}

\item{meta}{Data frame containing metadata, with SampleID in the first column and Group label in the second column}

\item{colors1}{Vector of colors for samples}

\item{colors2}{Vector of colors for clusters}
}
\value{
A ggplot object
}
\description{
Plot a sample based cluster for JoesFlow
}
